app: reinstate API wrapper to hide client implementation
