+++ app/proguard-rules.pro
@@ -8,3 +8,11 @@
 -keepclasseswithmembers class dev.msfjarvis.lobsters.model.** {
     kotlinx.serialization.KSerializer serializer(...);
 }
+
+# Workaround for https://github.com/ktorio/ktor/issues/1354
+-keepclassmembers class io.ktor.** { volatile <fields>; }
+
+# Workaround for https://github.com/Kotlin/kotlinx.coroutines/issues/1564
+-keepclassmembers class kotlinx.** {
+    volatile <fields>;
+}
